Super small grammar/punctuation fixes, formatting
The syntax/markup fix is focused on having anything that is an application (a binary) have an _italics_ markup so that it appears in a way that is similar to what a heavier markup such as DocBook shows for e.g. <application>. Regardless of the traditional tech writing appearance, it's useful for conversion via e.g. AsciiDoc to have a subset of visual styles tied to a specific subset of technical meaning. Anyway, that's why I get all pedantic with the _italics_. :)